Move DefaultAccessType to the same file as TenantMapping
